Unpipe stdin from child before restarting. Fixes #235
Note that I have to use setTimeout (I tried process.nextTick but it was still too soon) to delay between disconnecting the stdin pipe and actually restarting the child.

I've tested this with a simple script that echos stdin, and used the "rs" restart command over and over and it avoids the two EPIPE and ECONNRESET errors. I'm not sure I can easily create automated tests for this, but I'll try at a later date...maybe.