Fix #355
Very sutble, but ensures that we kill the child *before* the stdin connection is closed. This is because (for this particular bug) the child was listening to stdin.end and self terminating, which meant that our SIGUSR2 signal never made it to the child, which results in nodemon not realising that the child should be restarted. Hopefully this helps others too!