test to demonstrate coffee issue with --args
If you try to run nodemon with a coffee file script and pass an
argument to your application, it is instead adding a --nodejs when there
are no node flags to pass along. This causes the command to fail because
the script name is now passed as a flag to node.

Repro:

  nodemon script.coffee --my-argument

Expected:

  coffee script.coffee --my-argument

Actual:

  coffee --nodejs script.coffee --my-argument

Related to #530